use 16 bits as an index into arrays

+++ engine.h
@@ -256,13 +256,13 @@ int32_t getceilzofslope(int16_t sectnum, int32_t dax, int32_t day);
 int32_t getflorzofslope(int16_t sectnum, int32_t dax, int32_t day);
 int_fast8_t clipinsideboxline(int32_t x, int32_t y, int32_t x1, int32_t y1, int32_t x2, int32_t y2, int32_t walldist);
 int_fast8_t inside(int32_t x, int32_t y, int16_t sectnum);
-void drawsprite(int32_t snum);
-int_fast8_t spritewallfront(spritetype __far* s, int32_t w);
+void drawsprite(int_fast16_t snum);
+int_fast8_t spritewallfront(spritetype __far* s, int_fast16_t w);
 void clearlongbuf(int32_t __far* s, int32_t c, size_t n);
-void maskwallscan(int32_t x1, int32_t x2, int16_t __far* uwal, int16_t __far* dwal, int32_t __far* swal, int32_t __far* lwal, int16_t globalpicnum);
-void transmaskwallscan(int32_t x1, int32_t x2, int16_t globalpicnum);
+void maskwallscan(int_fast16_t x1, int_fast16_t x2, int16_t __far* uwal, int16_t __far* dwal, int32_t __far* swal, int32_t __far* lwal, int16_t globalpicnum);
+void transmaskwallscan(int_fast16_t x1, int_fast16_t x2, int16_t globalpicnum);
 int32_t krecip(int32_t num);
-uint8_t owallmost(int16_t __far* mostbuf, int32_t w, int32_t z);
+uint8_t owallmost(int16_t __far* mostbuf, int_fast16_t w, int32_t z);
 void clearshortbuf(int16_t __far* s, int16_t c, size_t n);
 void qinterpolatedown16short(void __far* bufptr, int32_t num, int32_t val, int32_t add);
